Minor changes buggy
